LayoutTests/editing/selection/ios/hide-selection-in-hidden-contenteditable-frame-expected.txt


Verifies that selection UI is suppressed when the editable root is transparent. To manually test, focus the box above and verify that:

The caret is not shown.
Selection highlights are not shown.
The selection cannot be changed via gesture.
Caret rect after focus: (left = 0, top = 0, width = 0, height = 0)
Selection rects after selecting all:
Selection start grabber rect after selecting all: (left = 0, top = 0, width = 0, height = 0)
Selection end grabber rect after selecting all: (left = 0, top = 0, width = 0, height = 0)
Selection before tap: ([object Text]#333, [object Text]#333)
Selection after tap: ([object Text]#333, [object Text]#333)
